feat: 楽譜上を移動する進捗のラインが滑らかに動くようにする
- `ScoreSection.svelte` のカーソルを、各描画時のSVGの静的な線から独立した絶対配置のDOM要素に変更
- `requestAnimationFrame` と `playerState.playbackStartTimeMs` を用いて、各拍間の位置を連続的に補間して滑らかに移動させるように実装
- 両方の譜面を貫通するようにカーソルの高さを設定
- README.md と AGENTS.md にカーソルのアニメーション方針に関するドキュメントを追記
- テストファイル `home.spec.ts` を更新し、新しいカーソル要素がDOMに存在することを検証

Fixes #27
